easing code of measurement filter
